Check for "double" instead of "float"
See http://php.net/manual/en/function.gettype.php for more information.
"double" is in PHP for historical reasons.

+++ src/JsonMapper.php
@@ -397,7 +397,7 @@ class JsonMapper
             || $type == 'string'
             || $type == 'boolean' || $type == 'bool'
             || $type == 'integer' || $type == 'int'
-            || $type == 'float';
+            || $type == 'double';
     }
 
     /**
